+++ kotlin/android/app/src/main/java/dev/firezone/android/tunnel/TunnelLogger.kt
@@ -1,0 +1,5 @@
+package dev.firezone.android.tunnel
+
+object TunnelLogger {
+    external fun init()
+}
