+++ kotlin/android/app/src/main/java/dev/firezone/android/tunnel/TunnelSession.kt
@@ -1,6 +1,6 @@
-package dev.firezone.connlib
+package dev.firezone.android.tunnel
 
-object Session {
+object TunnelSession {
     external fun connect(fd: Int, controlPlaneUrl: String, token: String, callback: Any): Long
     external fun disconnect(session: Long): Boolean
 }
