--- rust/connlib/clients/android/connlib/src/main/java/dev/firezone/connlib/Logger.kt
@@ -1,5 +1,0 @@
-package dev.firezone.connlib
-
-object Logger {
-    external fun init()
-}
