--- rust/connlib/clients/android/connlib/src/test/java/dev/firezone/connlib/ConnlibTest.kt
@@ -1,9 +1,0 @@
-package dev.firezone.connlib
-
-import org.junit.Test
-
-import org.junit.Assert.*
-
-class LoggerTest {
-    // TODO
-}
