--- rust/connlib/clients/android/connlib/src/test/java/dev/firezone/connlib/VpnServiceTest.kt
@@ -1,10 +1,0 @@
-
-package dev.firezone.connlib
-
-import org.junit.Test
-
-import org.junit.Assert.*
-
-class VpnServiceTest {
-    // TODO
-}
